Conflicts:
	ontwerpproject/src/controller/ServerHandler.java

+++ ontwerpproject/src/model/Component.java
@@ -43,7 +43,8 @@ public abstract class Component {
 	 * @invariant model != null
 	 */
 	protected Connection conn;
-
+	
+	//prepared sql statements never null
 	protected PreparedStatement check;
 	protected PreparedStatement delete;
 	protected PreparedStatement insert;
@@ -51,6 +52,7 @@ public abstract class Component {
 
 	/**
 	 * The intelligence of this component, should never be null
+	 * @invariant != null
 	 */
 	protected Intelligence intel;
 
@@ -260,7 +262,7 @@ public abstract class Component {
 		insert.setString(1, Long.toString(newdate));
 		insert.setString(2, "H");
 		for (int i = 0; i < b.length; ++i) {
-			insert.setString(i + 3, Integer.toString(b[i] / 24));
+			insert.setString(i + 3, Long.toString(b[i] / 24));
 		}
 		insert.executeUpdate();
 		conn.commit();
@@ -287,7 +289,7 @@ public abstract class Component {
 		insert.setString(1, Long.toString(newdate));
 		insert.setString(2, "H");
 		for (int i = 0; i < b.length; ++i) {
-			insert.setString(i + 3, Integer.toString(b[i] / 60));
+			insert.setString(i + 3, Long.toString(b[i] / 60));
 		}
 		insert.executeUpdate();
 		conn.commit();
@@ -315,7 +317,7 @@ public abstract class Component {
 		insert.setString(1, Long.toString(newdate));
 		insert.setString(2, "M");
 		for (int i = 0; i < b.length; ++i) {
-			insert.setString(i + 3, Integer.toString(b[i] / a));
+			insert.setString(i + 3, Long.toString(b[i] / a));
 		}
 		insert.executeUpdate();
 		conn.commit();
